chore: update dependencies and add Dockerfile language support
Bumps tree-sitter to 0.26 (with required u32 cast in node_to_json),
plus tree-sitter-css 0.25, tree-sitter-scala 0.26, tree-sitter-lua 0.5,
and toml 1.1. Adds Dockerfile parsing via tree-sitter-dockerfile-updated
to complete the filename-fallback support that was previously stubbed
out, restoring the failing test_table_alignment test.

+++ src/json.rs
@@ -13,7 +13,7 @@ pub(crate) struct JsonNode {
 
 pub(crate) fn node_to_json(source_code: &str, node: Node) -> JsonNode {
     let mut children = Vec::new();
-    for i in 0..node.child_count() {
+    for i in 0..node.child_count() as u32 {
         children.push(node_to_json(source_code, node.child(i).unwrap()));
     }
     let text_value = source_code[node.start_byte()..node.end_byte()].to_string();
